+++ src/sass/functions/macros.rs
@@ -19,18 +19,20 @@ macro_rules! one_arg {
 
 macro_rules! func {
     (( $($arg:ident $( = $value:expr )* ),* ), $body:expr) => {{
+        use crate::sass::Function;
         use std::sync::Arc;
-        SassFunction::builtin(vec![ $( one_arg!($arg $( = $value)* ) ),* ],
-                              false,
-                              Arc::new($body))
+        Function::builtin(vec![ $( one_arg!($arg $( = $value)* ) ),* ],
+                          false,
+                          Arc::new($body))
     }};
 }
 macro_rules! func_va {
     (( $($arg:ident $( = $value:expr )* ),* ), $body:expr) => {{
+        use crate::sass::Function;
         use std::sync::Arc;
-        SassFunction::builtin(vec![ $( one_arg!($arg $( = $value)* ) ),* ],
-                              true,
-                              Arc::new($body))
+        Function::builtin(vec![ $( one_arg!($arg $( = $value)* ) ),* ],
+                          true,
+                          Arc::new($body))
     }};
 }
 macro_rules! def {