+++ src/sass/mod.rs
@@ -1,6 +1,7 @@
 //! Value and Item types (and some supporting) for sass.
 mod call_args;
 mod formal_args;
+mod functions;
 mod item;
 mod mixin;
 mod name;
@@ -9,6 +10,7 @@ mod value;
 
 pub use self::call_args::CallArgs;
 pub use self::formal_args::FormalArgs;
+pub use self::functions::{get_global_module, Function};
 pub use self::item::{Item, UseAs};
 pub use self::mixin::Mixin;
 pub use self::name::Name;