+++ CHANGELOG.md
@@ -19,6 +19,7 @@ project adheres to
 * Some types moved into `value` mod, imporoved docs. PR #90.
 * Add support for `$=`, `~=` and `^=` in css attribute selectors. PR #92.
 * A `Numeric` is a Number with a Unit.  Also hide NumValue enum. PR #93.
+* The `SassFunction` type is now named `sass::Function`.
 * Test suite sass-spec updated to 2021-02-04.
 * Some cleanups.